+++ tests/Classes/DataPopulator.php
@@ -7,6 +7,8 @@ use Luma\AuroraDatabase\DatabaseConnection;
 use Luma\AuroraDatabase\Model\Aurora;
 use Luma\AuroraDatabase\Utils\Collection;
 use Luma\SecurityComponent\Authentication\Password;
+use Luma\SecurityComponent\Entity\Permission;
+use Luma\SecurityComponent\Entity\Role;
 
 class DataPopulator
 {
@@ -22,6 +24,8 @@ class DataPopulator
      * @param array $data
      *
      * @return void
+     *
+     * @throws \ReflectionException
      */
     public function populate(array $data): void
     {
@@ -58,6 +62,8 @@ class DataPopulator
      * @param array $roles
      *
      * @return void
+     *
+     * @throws \ReflectionException
      */
     private function saveUser(string $username, string $emailAddress, string $password, array $roles): void
     {