+++ tests/data/test.sql
@@ -1,15 +1,13 @@
-CREATE SCHEMA IF NOT EXISTS SecurityComponentTest;
-CREATE SCHEMA IF NOT EXISTS Security;
+DROP SCHEMA IF EXISTS Security;
+DROP SCHEMA IF EXISTS SecurityComponentTest;
+
+CREATE SCHEMA Security;
 
 USE Security;
 
-DROP TABLE IF EXISTS tblRoleUser;
+DROP TABLE IF EXISTS tblRoleUser, tblPermissionRole, ublRole, ublPermission;
 
-USE SecurityComponentTest;
-
-DROP TABLE IF EXISTS User, AddressDetails;
-
-CREATE TABLE User (
+CREATE TABLE tblUser (
     intUserId INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
     strUsername VARCHAR(60) NOT NULL,
     strEmailAddress VARCHAR(255) NOT NULL,
@@ -18,27 +16,23 @@ CREATE TABLE User (
     UNIQUE KEY (strEmailAddress)
 );
 
-CREATE TABLE AddressDetails (
+CREATE TABLE tblAddressDetails (
     intAddressDetailsId INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     strAddressLineOne VARCHAR(255) NOT NULL,
     strAddressLineTwo VARCHAR(255),
     strCity VARCHAR(60) NOT NULL,
     strPostcode VARCHAR(9) NOT NULL,
     intUserId INT(11) UNSIGNED NOT NULL,
-    FOREIGN KEY (intUserId) REFERENCES User(intUserId)
+    FOREIGN KEY (intUserId) REFERENCES tblUser(intUserId)
 );
 
-USE Security;
-
-DROP TABLE IF EXISTS tblPermissionRole, tblRole, tblPermission;
-
-CREATE TABLE tblRole (
+CREATE TABLE ublRole (
     intRoleId INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
     strRoleName VARCHAR(110) NOT NULL UNIQUE,
     strRoleHandle VARCHAR(110) NOT NULL UNIQUE
 );
 
-CREATE TABLE tblPermission (
+CREATE TABLE ublPermission (
     intPermissionId INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
     strPermissionName VARCHAR(110) NOT NULL UNIQUE,
     strPermissionHandle VARCHAR(110) NOT NULL UNIQUE
@@ -47,15 +41,15 @@ CREATE TABLE tblPermission (
 CREATE TABLE tblPermissionRole (
     intPermissionId INT(11) UNSIGNED NOT NULL,
     intRoleId INT(11) UNSIGNED NOT NULL,
-    FOREIGN KEY (intPermissionId) REFERENCES tblPermission(intPermissionId),
-    FOREIGN KEY (intRoleId) REFERENCES tblRole(intRoleId),
+    FOREIGN KEY (intPermissionId) REFERENCES ublPermission(intPermissionId),
+    FOREIGN KEY (intRoleId) REFERENCES ublRole(intRoleId),
     UNIQUE KEY (intPermissionId, intRoleId)
 );
 
 CREATE TABLE tblRoleUser (
     intUserId INT(11) UNSIGNED NOT NULL,
     intRoleId INT(11) UNSIGNED NOT NULL,
-    FOREIGN KEY (intUserId) REFERENCES SecurityComponentTest.user(intUserId),
-    FOREIGN KEY (intRoleId) REFERENCES tblRole(intRoleId),
+    FOREIGN KEY (intUserId) REFERENCES tblUser(intUserId),
+    FOREIGN KEY (intRoleId) REFERENCES ublRole(intRoleId),
     UNIQUE KEY (intUserId, intRoleId)
 );